# server/docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_ix-pay_ixpay-pro_internal_dto_base_request.CreateAPIRequest:
    properties:
      authRequired:
        type: boolean
      authType:
        enum:
        - 0
        - 1
        type: integer
      btnPermIds:
        items:
          type: string
        type: array
      description:
        maxLength: 500
        type: string
      group:
        maxLength: 100
        type: string
      menuIds:
        items:
          type: string
        type: array
      method:
        enum:
        - GET
        - POST
        - PUT
        - DELETE
        - PATCH
        - HEAD
        - OPTIONS
        type: string
      path:
        maxLength: 255
        type: string
      roleIds:
        items:
          type: string
        type: array
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - method
    - path
    type: object
...
  github_com_ix-pay_ixpay-pro_internal_dto_base_request.UpdateAPIRequest:
    properties:
      authRequired:
        type: boolean
      authType:
        enum:
        - 0
        - 1
        type: integer
      btnPermIds:
        items:
          type: string
        type: array
      description:
        maxLength: 500
        type: string
      group:
        maxLength: 100
        type: string
      id:
        type: string
      menuIds:
        items:
          type: string
        type: array
      method:
        enum:
        - GET
        - POST
        - PUT
        - DELETE
        - PATCH
        - HEAD
        - OPTIONS
        type: string
      path:
        maxLength: 255
        type: string
      roleIds:
        items:
          type: string
        type: array
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - id
    - method
    - path
    type: object
...
  github_com_ix-pay_ixpay-pro_internal_dto_base_response.APIResponse:
    properties:
      authRequired:
        type: boolean
      authType:
        type: integer
      btnPermIds:
        items:
          type: string
        type: array
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      group:
        type: string
      id:
        type: string
      menuIds:
        items:
          type: string
        type: array
      method:
        type: string
      path:
        type: string
      roleIds:
        items:
          type: string
        type: array
      status:
        type: integer
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
...
  github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: 127.0.0.1:8586
info:
  contact:
    email: support@example.com
    name: API Support
  description: 基于Gin和PostgreSQL的微信支付API服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 微信支付API服务
  version: "1.0"
paths:
...
  /api/admin/apis:
    get:
      consumes:
      - application/json
      description: 获取系统中注册的所有 API 路由信息（管理员权限），支持分页和搜索
      parameters:
      - description: 页码，默认为 1
        in: query
        name: page
        type: integer
      - description: 每页数量，默认为 100
        in: query
        name: pageSize
        type: integer
      - description: 搜索关键词（支持路径、描述、分组）
        in: query
        name: keyword
        type: string
      - description: 路由分组
        in: query
        name: group
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 路由列表
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.PageResult'
              type: object
        "401":
          description: 未授权
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "403":
          description: 无权限
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: 获取所有 API 路由列表
      tags:
      - API 路由管理
    post:
      consumes:
      - application/json
      description: 创建新的 API 路由信息（管理员权限）
      parameters:
      - description: API 路由信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_request.CreateAPIRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功的路由信息
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_response.APIResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "401":
          description: 未授权
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "403":
          description: 无权限
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: 创建 API 路由
      tags:
      - API 路由管理
  /api/admin/apis/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的 API 路由信息（管理员权限）
      parameters:
      - description: 路由 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "400":
          description: 参数错误
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "401":
          description: 未授权
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "403":
          description: 无权限
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "404":
          description: 路由不存在
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: 删除 API 路由
      tags:
      - API 路由管理
    get:
      consumes:
      - application/json
      description: 根据 ID 获取指定的 API 路由信息（管理员权限）
      parameters:
      - description: 路由 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 路由信息
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_response.APIResponse'
              type: object
        "401":
          description: 未授权
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "403":
          description: 无权限
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "404":
          description: 路由不存在
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: 根据 ID 获取 API 路由
      tags:
      - API 路由管理
    put:
      consumes:
      - application/json
      description: 更新指定的 API 路由信息（管理员权限）
      parameters:
      - description: 路由 ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新的路由信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_request.UpdateAPIRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功的路由信息
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_response.APIResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "401":
          description: 未授权
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "403":
          description: 无权限
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "404":
          description: 路由不存在
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: 更新 API 路由
      tags:
      - API 路由管理
  /api/admin/apis/list:
    get:
      consumes:
      - application/json
      description: 分页获取 API 路由列表，支持过滤（管理员权限）
      parameters:
      - description: 页码，默认为 1
        in: query
        name: page
        type: integer
      - description: 每页数量，默认为 10
        in: query
        name: pageSize
        type: integer
      - description: 路由分组
        in: query
        name: group
        type: string
      - description: 是否需要认证
        in: query
        name: authRequired
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 路由列表
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.PageResult'
                msg:
                  type: string
              type: object
        "401":
          description: 未授权
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "403":
          description: 无权限
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: 获取 API 路由列表
      tags:
      - API 路由管理
...
securityDefinitions:
  BearerAuth:
    in: header [Bearer ]
    name: Authorization
    type: apiKey
swagger: "2.0"